Marking mainly deprecation warnings as pure which allows for better tree-shaking on them, more can be read here https://iamakulov.com/notes/polished-webpack/
